refactored and moved Feed into separate file, reversed order of posts so most recent shows first
